Creating wiki page o!m modding tutorial
I wrote a modding tutorial a few weeks ago. While having it in the Forum might be good, a wiki link can be even better because suggestions from outside to add/edit stuff can be handled here better.

Hope that helps to improve the modding quality in o!m and maybe we could get even more BNs with that in the future!